Move status related code into separate file
