+++ components/telemetry-operator/controller/logpipeline/status_test.go
@@ -1,0 +1,1 @@
+package logpipeline
